requirements.txt
requests
python-dotenv
websocket-client 